requirements.txt
gspread
google-auth
requests
beautifulsoup4
feedparser